Move the uploadToLocal function to a separate module to support and encourage future code implementation.
